do-an-chuyen-nganh/services/receipt-ai/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
python-multipart
httpx
veryfi
